Done With the whole thing, I gave upgit add .!
it would be the worst thing I ever do, but that is it, I am done! The performance are going to be crazy and I will need to spend useless hours doing such a useless edits. So, I will probably archive this repo and create a new one starting it with react three fiber!